+++ runtime/litmus/src/migration/mod.rs
@@ -6,10 +6,10 @@ pub mod migrate_proxy;
 pub use migrate_proxy::ReplacePalletProxyStorage;
 pub mod migrate_vesting;
 pub use migrate_vesting::ReplacePalletVestingStorage;
-pub mod bridge_related;
-pub use bridge_related::ReplaceBridgeRelatedStorage;
-pub mod balances_transaction_payment;
-pub use balances_transaction_payment::ReplaceBalancesRelatedStorage;
+pub mod migrate_bridge_related;
+pub use migrate_bridge_related::ReplaceBridgeRelatedStorage;
+pub mod migrate_balances_transaction_payment;
+pub use migrate_balances_transaction_payment::ReplaceBalancesRelatedStorage;
 pub mod migrate_bounty;
 pub use migrate_bounty::ReplacePalletBountyStorage;
 pub mod migrate_democracy;
