+++ runtime/rococo/src/migration/migration.md
@@ -5,8 +5,11 @@ https://github.com/litentry/litentry-parachain/releases/tag/v0.9.19
 (1) token decimal change from 12 to 18
 (2) New token bridge related pallet storage migration.
 
+Rococo Only:
 For rococo, the initial migration contains some error code which make account.frozen = account.reserve * 10^6
 This storage error should be fixed in the following migration of rococo
+
+And the migration of parachain_staking, pallet_balances and bridge related migration also applied in this migration for rococo.
 
 # MigrateCollatorSelectionIntoParachainStaking
 P9100.rs
